nq-gnns/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=43.0.0", "wheel"]

[project]
dependencies = [
  "pytorch-lightning>=1.8.3",
  "torchmetrics==0.11.0",
  "hydra-core>=1.3.1",
  "hydra-colorlog>=1.2.0",
  "rich",
  "omegaconf",
  "tqdm",
  "numpy",
  "pandas==1.5.3",
  "seaborn",
  "scikit-learn",
  "matplotlib",
  "ipython",
  "jupyter",
  "pynvml",
  "pyrootutils",
  "h5py",
  "quapy",
]
name = "netquant"
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.0"

[project.scripts]
netquant-train = "netquant.train:main"
netquant-quant = "netquant.quantify:main"

[tool.setuptools.packages.find]
include = ["netquant*"] # package names should match these glob patterns (["*"] by default)
where = ["src"] # list of folders that contain the packages (["."] by default)

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
